requirements.txt
google-genai==0.3.0
python-dotenv==1.0.1
flask==3.0.3
gunicorn==23.0.0
Flask-SQLAlchemy==3.1.1
